// types.ts
export type QRContentType = 'link' | 'text' | 'vcard' | 'wifi' | 'clientPortal' | 'freelancerPortal' | 'publicForm';

export interface QRCodeRecord {
    id: string;
    name: string;
    type: QRContentType;
    content: string | Record<string, string>; // Can be a string or an object for complex types like vCard
}